Cancel statement if token cancellation is requested (#288)
* Change request behaviour

We want to get the response back from the cluster as fast as possible. When we have the statement_id from the response, we can cancel the statement if the cancellation token is cancelled.

We aim to minimize the time where user-code can't cancel the interaction with databricks cluster.

* Enhance comment to why no cancellation token is provided

* Use same exception types.
If the cancellation token is canceled within a task it raises a TaskCanceledException. Catch this exception and wrap it in an OperationCanceledException

* Bump package version

* Update release notes

* Update source/Databricks/source/SqlStatementExecution/Statement/DatabricksStatementRequest.cs

Co-authored-by: Dan Stenrøjl <51327761+dstenroejl@users.noreply.github.com>

* Fix cancellation of Databricks statements

* Add cancellation token test and helper classes

* Update release notes.
Using asynchronous mode for communication with job cluster

* Refactor method names

---------

Co-authored-by: Dan Stenrøjl <51327761+dstenroejl@users.noreply.github.com>